A custom hook has been created to extract responsive logic in the APISection component
