+++ src/components/APISection.jsx
@@ -1,4 +1,4 @@
-import { useEffect, useState } from 'react'
+import { useResponsive } from '@/hooks/useResponsive'
 import Image from 'next/image'
 import { Tab } from '@headlessui/react'
 import clsx from 'clsx'
@@ -12,22 +12,7 @@ import { API_FEATURES_LIST } from '@/constants'
 import { CopyAttributionButton } from './CopyAttributionButton'
 
 export function APISection () {
-  const [tabOrientation, setTabOrientation] = useState('horizontal')
-
-  useEffect(() => {
-    const lgMediaQuery = window.matchMedia('(min-width: 1024px)')
-
-    function onMediaQueryChange ({ matches }) {
-      setTabOrientation(matches ? 'vertical' : 'horizontal')
-    }
-
-    onMediaQueryChange(lgMediaQuery)
-    lgMediaQuery.addEventListener('change', onMediaQueryChange)
-
-    return () => {
-      lgMediaQuery.removeEventListener('change', onMediaQueryChange)
-    }
-  }, [])
+  const { tabOrientation } = useResponsive()
 
   return (
     <section
