+++ src/hooks/useResponsive.js
@@ -1,0 +1,22 @@
+import { useEffect, useState } from 'react'
+
+export const useResponsive = () => {
+  const [tabOrientation, setTabOrientation] = useState('horizontal')
+
+  useEffect(() => {
+    const lgMediaQuery = window.matchMedia('(min-width: 1024px)')
+
+    function onMediaQueryChange ({ matches }) {
+      setTabOrientation(matches ? 'vertical' : 'horizontal')
+    }
+
+    onMediaQueryChange(lgMediaQuery)
+    lgMediaQuery.addEventListener('change', onMediaQueryChange)
+
+    return () => {
+      lgMediaQuery.removeEventListener('change', onMediaQueryChange)
+    }
+  }, [])
+
+  return { tabOrientation }
+}
